backend/strategy-service/requirements.txt
TA-Lib
websockets
requests
redis
redis-om
python-binance
binance-futures-connector
grpcio
grpcio-tools